requirements.txt
langchain
langchain-community
langchainhub
google-generativeai
beautifulsoup4
requests
streamlit
tqdm
python-dotenv
faiss-cpu
langchain-openai
langchain-google-ai
streamlit-chat